test: move unit tests to test_super_component
