+++ test/core/super_component/test_super_component.py
@@ -16,6 +16,8 @@ from haystack.document_stores.types import DuplicatePolicy
 from haystack.utils.auth import Secret
 from haystack.core.serialization import default_from_dict, default_to_dict
 from haystack.core.super_component.super_component import InvalidMappingTypeError, InvalidMappingValueError
+from haystack.testing.sample_components import AddFixedValue, Double
+from unittest.mock import patch
 
 
 @pytest.fixture
@@ -107,6 +109,17 @@ def async_rag_pipeline(document_store):
     pipeline.connect("retriever.documents", "joiner.documents")
 
     return pipeline
+
+
+@pytest.fixture
+def sample_super_component():
+    """Creates a sample SuperComponent for testing visualization methods"""
+    pipe = Pipeline()
+    pipe.add_component("comp1", AddFixedValue(add=3))
+    pipe.add_component("comp2", Double())
+    pipe.connect("comp1.result", "comp2.value")
+
+    return SuperComponent(pipeline=pipe)
 
 
 @super_component
@@ -314,3 +327,42 @@ class TestSuperComponent:
         assert isinstance(custom_super_component, CustomSuperComponent)
         assert custom_super_component.var1 == 1
         assert custom_super_component.var2 == "test"
+
+    @patch("haystack.core.pipeline.Pipeline.show")
+    def test_show_delegates_to_pipeline(self, mock_show, sample_super_component):
+        """Test that SuperComponent.show() correctly delegates to Pipeline.show() with all parameters"""
+
+        server_url = "https://custom.mermaid.server"
+        params = {"theme": "dark", "format": "svg"}
+        timeout = 60
+
+        sample_super_component.show(server_url=server_url, params=params, timeout=timeout)
+        mock_show.assert_called_once_with(server_url=server_url, params=params, timeout=timeout)
+
+    @patch("haystack.core.pipeline.Pipeline.draw")
+    def test_draw_delegates_to_pipeline(self, mock_draw, sample_super_component, tmp_path):
+        """Test that SuperComponent.draw() correctly delegates to Pipeline.draw() with all parameters"""
+
+        path = tmp_path / "test_pipeline.png"
+        server_url = "https://custom.mermaid.server"
+        params = {"theme": "dark", "format": "png"}
+        timeout = 60
+
+        sample_super_component.draw(path=path, server_url=server_url, params=params, timeout=timeout)
+        mock_draw.assert_called_once_with(path=path, server_url=server_url, params=params, timeout=timeout)
+
+    @patch("haystack.core.pipeline.Pipeline.show")
+    def test_show_with_default_parameters(self, mock_show, sample_super_component):
+        """Test that SuperComponent.show() works with default parameters"""
+
+        sample_super_component.show()
+        mock_show.assert_called_once_with(server_url="https://mermaid.ink", params=None, timeout=30)
+
+    @patch("haystack.core.pipeline.Pipeline.draw")
+    def test_draw_with_default_parameters(self, mock_draw, sample_super_component, tmp_path):
+        """Test that SuperComponent.draw() works with default parameters except path"""
+
+        path = tmp_path / "test_pipeline.png"
+
+        sample_super_component.draw(path=path)
+        mock_draw.assert_called_once_with(path=path, server_url="https://mermaid.ink", params=None, timeout=30)
